Draw frigate holes with thin dashed lines in the system graph
The KineticJS parameters for 'dashArray' and 'dashArrayEnabled'
were changed to 'dash' and 'dashEnabled' respectively in v5.1.0
For some reason the CloudFlare CDN serves v5.1.9 when we ask for
v5.0.6., so these settings are currently not working.

Also, as far as I can see the 'eol' member variable is never set
on the connection nodes (GraphWormhole objects), so it wouldn't
trigger anyway.

I therefore updated the parameters, and changed to draw thin dashed
lines for frigate holes instead since we've had requests for that.